+++ src/power_specials.c
@@ -116,7 +116,7 @@ TbBool activate_bonus_level(struct PlayerInfo *player)
   LevelNumber sp_lvnum = get_loaded_level_number();
   TbBool result = set_bonus_level_visibility_for_singleplayer_level(player, sp_lvnum, true);
   if (!result)
-    ERRORLOG("No Bonus level assigned to level %d",(int)sp_lvnum);
+    ERRORLOG("No Bonus level assigned to level %ld",sp_lvnum);
   clear_flag(game.operation_flags, GOF_SingleLevel);
   return result;
 }
